With multiple vhosts on the same port sharing the same servername only one will be selected to match.
See http://httpd.apache.org/docs/2.4/vhosts/examples.html : "That means that if a request is received that does not match one of the specified ServerName directives, it will be served by this first VirtualHost." Using "localhost" for both will not solve the problem either.

Just converted to one vhost to keep it simple. We do not make changes on a finer grained level than the cds-apache docker container anyway.